requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.5.0
ecdsa==0.19.0
exceptiongroup==1.2.0
fastapi==0.110.1
h11==0.14.0
idna==3.6
iniconfig==2.0.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.0
pydantic==2.6.4
pydantic_core==2.16.3
PySocks==1.7.1
pytest==8.1.1
pytest-cov==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
rsa==4.9
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
tomli==2.0.1
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
webdriver-manager==4.0.1
wsproto==1.2.0
